docs:Tooltip rename#33
